[test] AndroidTest: move sample data
